+++ table/table_test.cc
@@ -7,6 +7,9 @@
 // Use of this source code is governed by a BSD-style license that can be
 // found in the LICENSE file. See the AUTHORS file for names of contributors.
 
+#include "rocksdb/table.h"
+
+#include <gtest/gtest.h>
 #include <stddef.h>
 #include <stdio.h>
 
@@ -26,8 +29,10 @@
 #include "memtable/stl_wrappers.h"
 #include "meta_blocks.h"
 #include "monitoring/statistics.h"
+#include "options/options_helper.h"
 #include "port/port.h"
 #include "rocksdb/cache.h"
+#include "rocksdb/compression_type.h"
 #include "rocksdb/db.h"
 #include "rocksdb/env.h"
 #include "rocksdb/file_checksum.h"
@@ -2193,6 +2198,115 @@ TEST_P(BlockBasedTableTest, SkipPrefixBloomFilter) {
   }
 }
 
+TEST_P(BlockBasedTableTest, BadChecksumType) {
+  BlockBasedTableOptions table_options = GetBlockBasedTableOptions();
+
+  Options options;
+  options.comparator = BytewiseComparator();
+  options.table_factory.reset(new BlockBasedTableFactory(table_options));
+
+  TableConstructor c(options.comparator);
+  InternalKey key("abc", 1, kTypeValue);
+  c.Add(key.Encode().ToString(), "test");
+  std::vector<std::string> keys;
+  stl_wrappers::KVMap kvmap;
+  const ImmutableOptions ioptions(options);
+  const MutableCFOptions moptions(options);
+  const InternalKeyComparator internal_comparator(options.comparator);
+  c.Finish(options, ioptions, moptions, table_options, internal_comparator,
+           &keys, &kvmap);
+
+  // Corrupt checksum type (123 is invalid)
+  auto& sink = *c.TEST_GetSink();
+  size_t len = sink.contents_.size();
+  ASSERT_EQ(sink.contents_[len - Footer::kNewVersionsEncodedLength], kCRC32c);
+  sink.contents_[len - Footer::kNewVersionsEncodedLength] = char{123};
+
+  // (Re-)Open table file with bad checksum type
+  const ImmutableOptions new_ioptions(options);
+  const MutableCFOptions new_moptions(options);
+  Status s = c.Reopen(new_ioptions, new_moptions);
+  ASSERT_NOK(s);
+  ASSERT_MATCHES_REGEX(s.ToString(), "Corruption: unknown checksum type 123.*");
+}
+
+namespace {
+std::string TrailerAsString(const std::string& contents,
+                            CompressionType compression_type,
+                            ChecksumType checksum_type) {
+  std::array<char, kBlockTrailerSize> trailer;
+  BlockBasedTableBuilder::ComputeBlockTrailer(contents, compression_type,
+                                              checksum_type, &trailer);
+  return Slice(trailer.data(), trailer.size()).ToString(/*hex*/ true);
+}
+}  // namespace
+
+// Make sure that checksum values don't change in later versions, even if
+// consistent within current version. (Other tests check for consistency
+// between written checksums and read-time validation, so here we only
+// have to verify the writer side.)
+TEST_P(BlockBasedTableTest, ChecksumSchemas) {
+  std::string b1 = "This is a short block!";
+  std::string b2;
+  for (int i = 0; i < 100; ++i) {
+    b2.append("This is a long block!");
+  }
+  CompressionType ct1 = kNoCompression;
+  CompressionType ct2 = kSnappyCompression;
+  CompressionType ct3 = kZSTD;
+
+  // Note: first byte of trailer is compression type, last 4 are checksum
+
+  for (ChecksumType t : GetSupportedChecksums()) {
+    switch (t) {
+      case kNoChecksum:
+        EXPECT_EQ(TrailerAsString(b1, ct1, t), "0000000000");
+        EXPECT_EQ(TrailerAsString(b1, ct2, t), "0100000000");
+        EXPECT_EQ(TrailerAsString(b1, ct3, t), "0700000000");
+        EXPECT_EQ(TrailerAsString(b2, ct1, t), "0000000000");
+        EXPECT_EQ(TrailerAsString(b2, ct2, t), "0100000000");
+        EXPECT_EQ(TrailerAsString(b2, ct3, t), "0700000000");
+        break;
+      case kCRC32c:
+        EXPECT_EQ(TrailerAsString(b1, ct1, t), "00583F0355");
+        EXPECT_EQ(TrailerAsString(b1, ct2, t), "012F9B0A57");
+        EXPECT_EQ(TrailerAsString(b1, ct3, t), "07ECE7DA1D");
+        EXPECT_EQ(TrailerAsString(b2, ct1, t), "00943EF0AB");
+        EXPECT_EQ(TrailerAsString(b2, ct2, t), "0143A2EDB1");
+        EXPECT_EQ(TrailerAsString(b2, ct3, t), "0700E53D63");
+        break;
+      case kxxHash:
+        EXPECT_EQ(TrailerAsString(b1, ct1, t), "004A2E5FB0");
+        EXPECT_EQ(TrailerAsString(b1, ct2, t), "010BD9F652");
+        EXPECT_EQ(TrailerAsString(b1, ct3, t), "07B4107E50");
+        EXPECT_EQ(TrailerAsString(b2, ct1, t), "0020F4D4BA");
+        EXPECT_EQ(TrailerAsString(b2, ct2, t), "018F1A1F99");
+        EXPECT_EQ(TrailerAsString(b2, ct3, t), "07A191A338");
+        break;
+      case kxxHash64:
+        EXPECT_EQ(TrailerAsString(b1, ct1, t), "00B74655EF");
+        EXPECT_EQ(TrailerAsString(b1, ct2, t), "01B6C8BBBE");
+        EXPECT_EQ(TrailerAsString(b1, ct3, t), "07AED9E3B4");
+        EXPECT_EQ(TrailerAsString(b2, ct1, t), "000D4999FE");
+        EXPECT_EQ(TrailerAsString(b2, ct2, t), "01F5932423");
+        EXPECT_EQ(TrailerAsString(b2, ct3, t), "076B31BAB1");
+        break;
+      case kXXH3:
+        EXPECT_EQ(TrailerAsString(b1, ct1, t), "00B37FB5E6");
+        EXPECT_EQ(TrailerAsString(b1, ct2, t), "016AFC258D");
+        EXPECT_EQ(TrailerAsString(b1, ct3, t), "075CE54616");
+        EXPECT_EQ(TrailerAsString(b2, ct1, t), "00FA2D482E");
+        EXPECT_EQ(TrailerAsString(b2, ct2, t), "0123AED845");
+        EXPECT_EQ(TrailerAsString(b2, ct3, t), "0715B7BBDE");
+        break;
+      default:
+        // Force this test to be updated on new ChecksumTypes
+        assert(false);
+        break;
+    }
+  }
+}
+
 void AddInternalKey(TableConstructor* c, const std::string& prefix,
                     std::string value = "v", int /*suffix_len*/ = 800) {
   static Random rnd(1023);
@@ -4036,40 +4150,20 @@ TEST(TableTest, FooterTests) {
     ASSERT_EQ(decoded_footer.index_handle().size(), index.size());
     ASSERT_EQ(decoded_footer.version(), 0U);
   }
-  {
-    // xxhash block based
+  for (auto t : GetSupportedChecksums()) {
+    // block based, various checksums
     std::string encoded;
     Footer footer(kBlockBasedTableMagicNumber, 1);
     BlockHandle meta_index(10, 5), index(20, 15);
     footer.set_metaindex_handle(meta_index);
     footer.set_index_handle(index);
-    footer.set_checksum(kxxHash);
+    footer.set_checksum(t);
     footer.EncodeTo(&encoded);
     Footer decoded_footer;
     Slice encoded_slice(encoded);
     ASSERT_OK(decoded_footer.DecodeFrom(&encoded_slice));
     ASSERT_EQ(decoded_footer.table_magic_number(), kBlockBasedTableMagicNumber);
-    ASSERT_EQ(decoded_footer.checksum(), kxxHash);
-    ASSERT_EQ(decoded_footer.metaindex_handle().offset(), meta_index.offset());
-    ASSERT_EQ(decoded_footer.metaindex_handle().size(), meta_index.size());
-    ASSERT_EQ(decoded_footer.index_handle().offset(), index.offset());
-    ASSERT_EQ(decoded_footer.index_handle().size(), index.size());
-    ASSERT_EQ(decoded_footer.version(), 1U);
-  }
-  {
-    // xxhash64 block based
-    std::string encoded;
-    Footer footer(kBlockBasedTableMagicNumber, 1);
-    BlockHandle meta_index(10, 5), index(20, 15);
-    footer.set_metaindex_handle(meta_index);
-    footer.set_index_handle(index);
-    footer.set_checksum(kxxHash64);
-    footer.EncodeTo(&encoded);
-    Footer decoded_footer;
-    Slice encoded_slice(encoded);
-    ASSERT_OK(decoded_footer.DecodeFrom(&encoded_slice));
-    ASSERT_EQ(decoded_footer.table_magic_number(), kBlockBasedTableMagicNumber);
-    ASSERT_EQ(decoded_footer.checksum(), kxxHash64);
+    ASSERT_EQ(decoded_footer.checksum(), t);
     ASSERT_EQ(decoded_footer.metaindex_handle().offset(), meta_index.offset());
     ASSERT_EQ(decoded_footer.metaindex_handle().size(), meta_index.size());
     ASSERT_EQ(decoded_footer.index_handle().offset(), index.offset());
@@ -4098,7 +4192,7 @@ TEST(TableTest, FooterTests) {
     ASSERT_EQ(decoded_footer.version(), 0U);
   }
   {
-    // xxhash block based
+    // xxhash plain table (not currently used)
     std::string encoded;
     Footer footer(kPlainTableMagicNumber, 1);
     BlockHandle meta_index(10, 5), index(20, 15);